+++ src/AvroConvert/Features/GenerateModel/GenerateModel.cs
@@ -1,4 +1,25 @@
-﻿using System;
+﻿#region license
+/**Copyright (c) 2021 Adrian Strugala
+*
+* Licensed under the CC BY-NC-SA 3.0 License(the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* https://creativecommons.org/licenses/by-nc-sa/3.0/
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+* You are free to use or modify the code for personal usage.
+* For commercial usage purchase the product at
+*
+* https://xabe.net/product/avroconvert/
+*/
+#endregion
+
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,17 +28,17 @@ using Newtonsoft.Json.Linq;
 using SolTechnology.Avro.AvroObjectServices.Schema.Abstract;
 using SolTechnology.Avro.Infrastructure.Exceptions;
 
-namespace SolTechnology.Avro.Features.GenerateClass
+namespace SolTechnology.Avro.Features.GenerateModel
 {
-    public class GenerateClassHandler
+    public class GenerateModel
     {
-        public string HandleAvroObject(byte[] avroData)
+        public string FromAvroObject(byte[] avroData)
         {
             string schema = AvroConvert.GetSchema(avroData);
-            return HandleAvroSchema(schema);
+            return FromAvroSchema(schema);
         }
 
-        public string HandleAvroSchema(string schema)
+        public string FromAvroSchema(string schema)
         {
             JObject json = (JObject)JsonConvert.DeserializeObject(schema);